fix(hub): merge Library into Knowledge tab, fix /api/library/tree 500
The /library page was failing with HTTP 500: "permission denied for
table knowledge_entries". Two issues:

1. The factorylm_app role used by withTenantContext lacked SELECT on
   knowledge_entries / kg_entities / kg_relationships. Migration 011
   grants those.
2. 003_kb_hardening.sql installed an RLS policy keyed off
   app.current_tenant_id, but withTenantContext only writes
   app.tenant_id. Once the GRANT was fixed this would have silently
   returned zero rows. Migration 011 refreshes the policy to read
   either key, and tenant-context.ts now writes both for back-compat
   with older migrations.

Mike also wants the library merged into the existing Knowledge tab
rather than living at /library:

- /knowledge becomes the single browse-and-ingest surface: upload
  picker + active uploads at top, manufacturer/model/document tree
  below, chunk-detail sheet on tap.
- /library now permanently redirects to /knowledge so saved links
  and QR codes still work.
- "Library" is removed from the sidebar nav and i18n.
- Layout uses the existing theme variables and is mobile-responsive
  (≥44 px tap targets, sticky search header, bottom-sheet detail).
