+++ mira-hub/src/messages/en.json
@@ -6,7 +6,6 @@
     "actions": "Actions",
     "alerts": "Alerts",
     "knowledge": "Knowledge",
-    "library": "Library",
     "workOrders": "Work Orders",
     "assets": "Assets",
     "channels": "Channels",
